Organise dashboard project structure
